requirements.txt
cycler==0.10.0
jedi==0.17.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
parso==0.7.1
Pillow==8.3.2
pip==20.2.3
pluggy==1.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
setuptools==49.2.1
six==1.16.0
ujson==4.2.0
